:truck: rename: mdx components
