+++ src/components/MDX/Atoms/Code/index.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useEffect, useRef, useState } from "react"
 
 import CodeCopyBtn from "./CodeCopyBtn"
