+++ src/components/MDX/Atoms/Code/useTimeout.ts
@@ -1,3 +1,5 @@
+"use client"
+
 import { useEffect } from "react"
 
 const useTimeout = ({
