+++ src/components/MDX/Bundler/index.tsx
@@ -1,10 +1,12 @@
+"use client"
+
 import { useMemo } from "react"
 import { getMDXComponent } from "mdx-bundler/client"
-import { MDXAtoms } from "../Atoms"
+import { MDXAtoms } from "../atoms"
+
 interface MDXBundlerProp {
     source: string
 }
-
 const MDXBundler = ({ source }: MDXBundlerProp) => {
     const BundledComponent = useMemo(() => getMDXComponent(source), [source])
